requirements.txt
# add this line (or tighten it if present)
httpx<0.28
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-dotenv==1.0.1
azure-storage-blob==12.22.0
azure-search-documents==11.6.0b6
openai==1.51.2
langchain==0.3.4
langchain-community==0.3.3
pydantic==2.9.2
pytesseract==0.3.13
Pillow==10.4.0
docx2txt==0.8
pypdf==5.0.1
tqdm==4.66.5
numpy==1.26.4
pytest==8.3.3
pandas==2.2.2
pyarrow==17.0.0
numpy==1.26.4

